Reorganize Job's members
